Added Image example
