+++ test/examples/CMakeLists.txt
@@ -12,5 +12,5 @@ add_executable(writing_complex writing_complex.cpp)
 target_link_libraries(writing_complex h5cpp_shared)
 
 add_executable(writing_image writing_image.cpp
-                             pixel.cpp)
+                             rgbpixel.cpp)
 target_link_libraries(writing_image h5cpp_shared)
