+++ test/examples/attributes.cpp
@@ -1,0 +1,10 @@
+/*
+ * attributes.cpp
+ *
+ *  Created on: Oct 7, 2017
+ *      Author: eugen
+ */
+
+
+
+
